requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
psycopg2==2.8.6  # https://github.com/psycopg/psycopg2

# Django
# ------------------------------------------------------------------------------
django-anymail==8.1  # https://github.com/anymail/django-anymail
django-background-tasks==1.2.5
beautifulsoup4==4.9.3
